+++ src/Controller/ChartsController.php
@@ -5,16 +5,28 @@ namespace App\Controller;
 use App\Entity\Charts;
 use App\Form\ChartsType;
 use App\Repository\ChartsRepository;
+use App\Service\UserVerificationService;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/charts')]
 class ChartsController extends AbstractController {
+    private $userVerificationService;
+
+    public function __construct(UserVerificationService $userVerificationService)
+    {
+        $this->userVerificationService = $userVerificationService;
+    }
+    
     #[Route('/', name: 'app_charts_index', methods: ['GET'])]
     public function index(ChartsRepository $chartsRepository, Request $request): Response {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         // $chartGlobal = new Charts(); // Créez une nouvelle instance de Charts (ou adaptez ceci selon vos besoins)
         return $this->render('charts/index.html.twig', [
             'charts' => $chartsRepository->findAll(),
@@ -23,6 +35,11 @@ class ChartsController extends AbstractController {
 
     #[Route('/new', name: 'app_charts_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response {
+
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $chart = new Charts();
         $form = $this->createForm(ChartsType::class, $chart);
         $form->handleRequest($request);
@@ -42,6 +59,11 @@ class ChartsController extends AbstractController {
 
     #[Route('/{id}', name: 'app_charts_show', methods: ['GET'])]
     public function show(Charts $chart): Response {
+
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('charts/show.html.twig', [
             'chart' => $chart,
         ]);
@@ -49,6 +71,11 @@ class ChartsController extends AbstractController {
 
     #[Route('/{id}/edit', name: 'app_charts_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Charts $chart, EntityManagerInterface $entityManager): Response {
+
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         $form = $this->createForm(ChartsType::class, $chart);
         $form->handleRequest($request);
 
@@ -66,6 +93,11 @@ class ChartsController extends AbstractController {
 
     #[Route('/{id}', name: 'app_charts_delete', methods: ['POST'])]
     public function delete(Request $request, Charts $chart, EntityManagerInterface $entityManager): Response {
+
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$chart->getId(), $request->request->get('_token'))) {
             $entityManager->remove($chart);
             $entityManager->flush();
