+++ src/Controller/PlayerController.php
@@ -5,18 +5,30 @@ namespace App\Controller;
 use App\Entity\Player;
 use App\Form\PlayerType;
 use App\Repository\PlayerRepository;
+use App\Service\UserVerificationService;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/player')]
 class PlayerController extends AbstractController
 {
+    private $userVerificationService;
+
+    public function __construct(UserVerificationService $userVerificationService)
+    {
+        $this->userVerificationService = $userVerificationService;
+    }
+    
     #[Route('/', name: 'app_player_index', methods: ['GET'])]
     public function index(PlayerRepository $playerRepository): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('player/index.html.twig', [
             'players' => $playerRepository->findAll(),
         ]);
@@ -25,6 +37,10 @@ class PlayerController extends AbstractController
     #[Route('/new', name: 'app_player_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $player = new Player();
         $form = $this->createForm(PlayerType::class, $player);
         $form->handleRequest($request);
@@ -45,6 +61,10 @@ class PlayerController extends AbstractController
     #[Route('/{id}', name: 'app_player_show', methods: ['GET'])]
     public function show(Player $player): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('player/show.html.twig', [
             'player' => $player,
         ]);
@@ -53,6 +73,10 @@ class PlayerController extends AbstractController
     #[Route('/{id}/edit', name: 'app_player_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Player $player, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         $form = $this->createForm(PlayerType::class, $player);
         $form->handleRequest($request);
 
@@ -71,6 +95,10 @@ class PlayerController extends AbstractController
     #[Route('/{id}', name: 'app_player_delete', methods: ['POST'])]
     public function delete(Request $request, Player $player, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$player->getId(), $request->request->get('_token'))) {
             $entityManager->remove($player);
             $entityManager->flush();
@@ -82,6 +110,10 @@ class PlayerController extends AbstractController
     #[Route('/{id}/poste', name: 'app_player_poste', methods: ['GET'])]
     public function poste(Player $player): Response
     {
+        if(!$this->userVerificationService->verifyUser()){
+            return $this->redirectToRoute('app_verif_code', [], Response::HTTP_SEE_OTHER);
+        }
+
         return $this->render('player/poste.html.twig', [
             'player' => $player,
         ]);
